chore(demo): refactor demo module structure and enhance functionality with new demo accounts and transactions
