+++ README.md
@@ -12,7 +12,11 @@ yarn serve
 
 If you need to interact with a real account and data, then you will also need to run the [Nimiq Hub](https://github.com/nimiq/hub#contribute) and [Nimiq Keyguard](https://github.com/nimiq/keyguard/#development) in development, too.
 
-Otherwise, you can activate demo mode by appending `?demo` to the URL: `http://localhost:8081/?demo`.
+For demo mode with simulated accounts and transactions, use:
+```
+yarn serve:demo
+```
+
 ### Compiles and minifies for production
 ```
 yarn build
